Split TestHooks files to one type per file
