+++ src/Workspaces/Core/Portable/Shared/TestHooks/AsynchronousOperationListener+AsyncToken.cs
