+++ src/Workspaces/Core/Portable/Shared/TestHooks/AsynchronousOperationListener+DiagnosticAsyncToken.cs
